Add files via upload
This is a heavily commented version of the Barrow climatology. 
This acts as a template to explain what is happening within the code, and select metadata to orient the user.

Use this script instead of the CLEANED_Barrow_Climatology.rmd if you aim to understand the reasoning behind the code better.